+++ std.go
@@ -325,6 +325,79 @@ npm.test = function test() {
   return npm.apply(null, ['test'].concat(Array.prototype.slice.call(arguments)));
 };
 `),
+	"language/python.js": []byte(`//
+// aster :: language/python.js
+//
+//   Copyright (c) 2018 Akinori Hattori <hattya@gmail.com>
+//
+//   Permission is hereby granted, free of charge, to any person
+//   obtaining a copy of this software and associated documentation files
+//   (the "Software"), to deal in the Software without restriction,
+//   including without limitation the rights to use, copy, modify, merge,
+//   publish, distribute, sublicense, and/or sell copies of the Software,
+//   and to permit persons to whom the Software is furnished to do so,
+//   subject to the following conditions:
+//
+//   The above copyright notice and this permission notice shall be
+//   included in all copies or substantial portions of the Software.
+//
+//   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+//   EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+//   MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+//   NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
+//   BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+//   ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+//   CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+//   SOFTWARE.
+//
+
+'use strict';
+
+var os = require('os');
+var language = require('language');
+
+var coverage = exports.coverage = function coverage() {
+  if (!os.whence('coverage')) {
+    aster.notify('failure', exports.prefix + 'coverage', 'coverage not found!');
+    return true;
+  }
+
+  return language.system({
+    args: ['coverage'].concat(Array.prototype.slice.call(arguments)),
+    title: 'coverage',
+    success: arguments[0] + ' passed',
+    failure: arguments[0] + ' failed',
+  });
+};
+
+coverage.annotate = function annotate() {
+  return coverage.apply(null, ['annotate'].concat(Array.prototype.slice.call(arguments)));
+};
+
+coverage.combine = function combine() {
+  return coverage.apply(null, ['combine'].concat(Array.prototype.slice.call(arguments)));
+};
+
+coverage.erase = function erase() {
+  return coverage.apply(null, ['erase'].concat(Array.prototype.slice.call(arguments)));
+};
+
+coverage.html = function html() {
+  return coverage.apply(null, ['html'].concat(Array.prototype.slice.call(arguments)));
+};
+
+coverage.report = function report() {
+  return coverage.apply(null, ['report'].concat(Array.prototype.slice.call(arguments)));
+};
+
+coverage.run = function run() {
+  return coverage.apply(null, ['run'].concat(Array.prototype.slice.call(arguments)));
+};
+
+coverage.xml = function xml() {
+  return coverage.apply(null, ['xml'].concat(Array.prototype.slice.call(arguments)));
+};
+`),
 	"language/restructuredtext.js": []byte(`//
 // aster :: language/restructuredtext.js
 //